requirements.txt
hydra-core >= 1.3
praw >= 7.7.0
rich >= 14.0.0
requests
feedparser

rally @ git+https://github.com/anton-pershin/rally.git
